+++ src/caselawclient/client_helpers/search_helpers.py
@@ -9,12 +9,10 @@ def search_judgments_and_parse_response(
     """
     Search for judgments using the given search parameters and parse the response into a SearchResponse object.
 
-    Args:
-        api_client (MarklogicApiClient): An instance of MarklogicApiClient used to make the search request.
-        search_parameters (SearchParameters): An instance of SearchParameters containing the search parameters.
+    :param api_client: An instance of MarklogicApiClient used to make the search request
+    :param search_parameters: An instance of SearchParameters containing the search parameters
 
-    Returns:
-        SearchResponse: The parsed search response as a SearchResponse object.
+    :return: The parsed search response as a SearchResponse object
     """
     return SearchResponse.from_response_string(
         api_client.search_judgments_and_decode_response(search_parameters)
@@ -27,12 +25,10 @@ def search_and_parse_response(
     """
     Search using the given search parameters and parse the response into a SearchResponse object.
 
-    Args:
-        api_client (MarklogicApiClient): An instance of MarklogicApiClient used to make the search request.
-        search_parameters (SearchParameters): An instance of SearchParameters containing the search parameters.
+    :param api_client: An instance of MarklogicApiClient used to make the search request
+    :search_parameters: An instance of SearchParameters containing the search parameters
 
-    Returns:
-        SearchResponse: The parsed search response as a SearchResponse object.
+    :return: The parsed search response as a SearchResponse object
     """
     return SearchResponse.from_response_string(
         api_client.search_and_decode_response(search_parameters)