+++ src/caselawclient/models/__init__.py
@@ -1,0 +1,3 @@
+"""
+Classes which represent distinct entities within MarkLogic.
+"""